Implement basic path tracing

+++ raytracer/inc/material.h
@@ -15,22 +15,22 @@ public:
     };
 
 private:
-    glm::vec3 _diffuse_color;
-    glm::vec3 _emission;
-    glm::vec4 _albedo;
-    float _specular_exponent;
-    float _refractive_index;
+    glm::dvec3 _diffuse_color;
+    glm::dvec3 _emission;
+    glm::dvec4 _albedo;
+    double _specular_exponent;
+    double _refractive_index;
     Type _type;
 
 public:
     Material() :
-        _diffuse_color(0.0f), _emission(0.0f),
-        _albedo(1.0f, 0.0f, 0.0f, 0.0f),
-        _specular_exponent(0.0f), _refractive_index(1.0f),
+        _diffuse_color(0.0d), _emission(0.0d),
+        _albedo(1.0d, 0.0d, 0.0d, 0.0d),
+        _specular_exponent(0.0d), _refractive_index(1.0d),
         _type(DIFFUSE) {}
-    Material(const glm::vec3 &diffuse_color, const glm::vec3 &emission,
-            const glm::vec4 &albedo,
-            float specular_exponent, float refractive_index,
+    Material(const glm::dvec3 &diffuse_color, const glm::dvec3 &emission,
+            const glm::dvec4 &albedo,
+            double specular_exponent, double refractive_index,
             Type type) :
         _diffuse_color(diffuse_color), _emission(emission),
         _albedo(albedo),
@@ -38,11 +38,11 @@ public:
         _refractive_index(refractive_index),
         _type(type) {}
 
-    const glm::vec3 &diffuse_color() const { return _diffuse_color; }
-    const glm::vec3 &emission() const { return _emission; }
-    const glm::vec4 &albedo() const { return _albedo; }
-    float specular_exponent() const { return _specular_exponent; }
-    float refractive_index() const { return _refractive_index; }
+    const glm::dvec3 &diffuse_color() const { return _diffuse_color; }
+    const glm::dvec3 &emission() const { return _emission; }
+    const glm::dvec4 &albedo() const { return _albedo; }
+    double specular_exponent() const { return _specular_exponent; }
+    double refractive_index() const { return _refractive_index; }
     Type type() const { return _type; }
 };
